% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mccxmpl}
\alias{mccxmpl}
\title{Example data for matched case-control analysis}
\format{
\subsection{\code{mccxmpl}}{

A data frame with 27 rows and 2 columns:
\describe{
\item{\code{case}}{Integer variable of either \code{1} (exposed) or \code{0}
(not exposed)}
\item{\code{control}}{Integer variable of either \code{1} (exposed) or \code{0}
(not exposed)}
}
}
}
\source{
In Stata 18 - run the commands:

\if{html}{\out{<div class="sourceCode">}}\preformatted{webuse mccxmpl, clear
expand pop
keep case control
}\if{html}{\out{</div>}}
}
\usage{
mccxmpl
}
\description{
A subset of data from Jick \emph{et al}. (1973) with data on a matched
case-control study on myocardial infarction and drinking 6+ cups of coffee
per day. Cases and controls were matched after excluding people who drank
1 to 5 cups of coffee per day.
}
\references{
Jick, H. \emph{et al}. (1973). Coffee and myocardial infarction. \strong{New England
Journal of Medicine} 289: 63–67. \doi{10.1056/NEJM197307122890203}.
}
\keyword{datasets}
